requirements.txt


Django
djangorestframework
redis
django-ckeditor
googletrans==4.0.0-rc1
